refactor(web): extract DashboardWidgets, replace window.confirm, use formatCents, drop last_used_at
H6: Move ActionQueue, JobsToday, Materials and their types (CountAction,
CommandVisit, MaterialJob) into a new DashboardWidgets.tsx. WorkdayPanel
and DailyCommandCenter re-export them for backward compatibility; direct
import sites (page.tsx, OwnerDashboard.tsx, DailyCommandCenter.tsx) updated
to point at the new file.

H1: Replace window.confirm in WorkdayPanel.discardSession with ConfirmDialog
component, matching the pattern used by DeleteJobButton.

H4: Replace three inline (cents/100).toLocaleString(...) calls in
WorkdayPanel's Financial Snapshot with formatCents from @/lib/money.
Remove the local dollars() helper in OwnerDashboard and use formatCents there
too.

L2: last_used_at subquery was already absent from HEAD in my-day/page.tsx;
working-tree copy restored to match.
